backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
alembic>=1.13.1
pydantic>=2.9.0
groq>=0.9.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
httpx>=0.27.0
python-dotenv>=1.0.1
pypdf>=4.2.0
python-docx>=1.1.2